+++ .github/workflows/ci.yml
@@ -12,7 +12,7 @@ jobs:
       - uses: actions/checkout@v3
       - uses: actions/setup-node@v4
         with:
-          node-version: '20'
+          node-version: '22'
       - name: Cache node_modules
         uses: actions/cache@v4
         env:
@@ -31,7 +31,7 @@ jobs:
       - uses: actions/checkout@v3
       - uses: actions/setup-node@v4
         with:
-          node-version: '20'
+          node-version: '22'
       - name: Cache node_modules
         uses: actions/cache@v4
         env:
@@ -97,7 +97,7 @@ jobs:
       - uses: actions/checkout@v3
       - uses: actions/setup-node@v4
         with:
-          node-version: '20'
+          node-version: '22'
       - name: Cache node_modules
         uses: actions/cache@v4
         env:
@@ -170,12 +170,17 @@ jobs:
             attempt=$((attempt + 1))
           done
 
-      - name: integration
+      - name: integration (HTTP)
         run: npm run test:integration:cover
         env:
           INDEXING_RETRY_INTERVAL: 4000
           INDEXING_MAX_RETRIES: 120
           NODE_URL: 'http://127.0.0.1:8001'
+
+      - name: integration (P2P)
+        run: npm run test:integration
+        env:
+          NODE_URL: '/ip4/172.15.0.5/tcp/9001/ws/p2p/16Uiu2HAmRkJeRYRghP3ETQCpdz8NsQzQE9RpSST7i5YNgWqH4dVE'
 
       - name: docker logs
         run: |
@@ -199,7 +204,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        node: ['20', '18']
+        node: ['22', '20']
 
     steps:
       - uses: actions/checkout@v3
@@ -237,7 +242,7 @@ jobs:
       - uses: actions/checkout@v3
       - uses: actions/setup-node@v4
         with:
-          node-version: '20'
+          node-version: '22'
 
       - name: checkout ocean.js repo
         uses: actions/checkout@v3
