+++ package.json
@@ -56,6 +56,11 @@
     "web3": "^1.8.0"
   },
   "dependencies": {
+    "@chainsafe/libp2p-noise": "^17.0.0",
+    "@chainsafe/libp2p-yamux": "^8.0.1",
+    "@libp2p/bootstrap": "^12.0.14",
+    "@libp2p/websockets": "^10.1.6",
+    "@multiformats/multiaddr": "^13.0.1",
     "@oasisprotocol/sapphire-paratime": "^1.3.2",
     "@oceanprotocol/contracts": "^2.6.0",
     "@oceanprotocol/ddo-js": "^0.1.4",
@@ -69,7 +74,9 @@
     "eciesjs": "^0.4.5",
     "ethers": "^6.15.0",
     "form-data": "^2.3.3",
-    "jsonwebtoken": "^9.0.2"
+    "jsonwebtoken": "^9.0.2",
+    "libp2p": "^3.1.6",
+    "uint8arrays": "^4.0.10"
   },
   "devDependencies": {
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
